+++ src/main/java/com/controle/financeiro/dto/PagamentoDto.java
@@ -1,7 +1,8 @@
 package com.controle.financeiro.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record PagamentoDto(double valorDoPagamento, LocalDate dataPagamento) {
+public record PagamentoDto(BigDecimal valorDoPagamento, LocalDate dataPagamento) {
     
 }
